Add files via upload
Version 2 of the models. Now contains the smaller model. Had to split out the minmax scalers to relabel variables. (Might be more useful for Saka in referencing them in a different notebook?) BE AWARE that scaling happens once for each model, as this only works after feature selection.
The process of dropping 'Infant_deaths' is now more explicit, based on vif scores.
There is some commenting, but still need to relate the notebook to Sian's feature engineering.